+++ src/email/ui/MailCopy.tsx
